Fix cutting tests
